+++ src/armstrong/src/plugins/midi/miditime.cpp
@@ -118,7 +118,7 @@ void miditimemachine::midi_tick() {
 }
 
 void miditimemachine::midi_out(int time, unsigned int data) {
-	zzub::midi_message msg = { -1, data, time };
+	zzub::midi_message msg = { -1, data, (unsigned long)time };
 	_mixer->midi_out(_id, msg);
 }
